Refactoring: Move some method from ServiceUtils
ReadFileToString moved to orbit_base, ReadModules and
ParseMaps moved to ElfUtils.

Test: run OrbitBaseTests and ElfUtilsTests
