+++ OrbitBase/ThreadPoolTest.cpp
@@ -2,11 +2,9 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include <gmock/gmock-matchers.h>
 #include <gtest/gtest.h>
 
 #include <memory>
-#include <utility>
 
 #include "OrbitBase/ThreadPool.h"
 #include "absl/synchronization/mutex.h"
